PM Update 1009 - Implementing MT5 Data to ABCD, cleaning up Data inverval checker + exporting summaries, rounding on signalHandler PL, bugs in Charting WB, LSTM signal Label model

+++ Charting Work/ABCD_mapping.py
@@ -133,6 +133,6 @@ def ABCD_mapping(dataWithZigZags, thresholdMean = 0, thresholdVar = np.inf, labe
         signals[longIdxs] = 1
         signals[shortIdxs] = -1
 
-        dataWithZigZagsABCD['Signal'] = signals
+        dataWithZigZagsABCD['Signal'] = signals.astype('int')
 
     return dataWithZigZagsABCD
